+++ FreeRTOS/Demo/Common/include/fileIO.h
@@ -68,7 +68,7 @@
 
 void vDisplayMessage( const char * const pcMessageToPrint );
 void vWriteMessageToDisk( const char * const pcMessage );
-void vWriteBufferToDisk( const char * const pcBuffer, unsigned long ulBufferLength );
+void vWriteBufferToDisk( const char * const pcBuffer, uint32_t ulBufferLength );
 
 #endif
